Adjusted styling of the card component
Added several adjustments in different places of the application. It shouldn't break anything, but it would be always a good idea to double check in the other pages to make sure everything looks as supposed. Also note that some elements have been modified to accept external css, it is not the optimal way, but I am not sure if anything else is in place to allow me to modify the default styling of a component from it's parent