[WIP] DO NOT MERGE!!!!
Added a bunch of different fixes.

<< Missing >>
Currently the pledge amount is a wrong number, currently trying to fix it.

<< Fixes >>

First: Fixed problem about creating a new Dare, I am now returning only the status of the request instead of the payload. Checking for the payload was going to be a more difficult implementation at this stage

Second: Adjusted the pledge schema adding the projectId key as a NON mandatory field. this should not cause any conflicts with the current implementations

Nnumber: Removed the getProjectId utility function, since it was missing in the first place form the project

Nnumber + 1: Removed un-necessary parameters left behind from code refactoring (probably). This has made the pledge project functionality work again

Nnumber + 2: Adjusted the pledgeDare test case to generate a random user when you create a new project, main reason is to avoid duplication on Dare pledging logic